+++ lexer.c
@@ -6,12 +6,11 @@
 /*   By: mde-arpe <mde-arpe@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/27 04:25:53 by mde-arpe          #+#    #+#             */
-/*   Updated: 2023/08/13 19:19:43 by mde-arpe         ###   ########.fr       */
+/*   Updated: 2023/08/13 19:45:33 by mde-arpe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "lexer.h"
-#include "utils.h"
 
 // returns NULL if any of these happen:
 // - malloc fail
