+++ minishell.h
@@ -6,7 +6,7 @@
 /*   By: mde-arpe <mde-arpe@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/26 03:48:31 by mde-arpe          #+#    #+#             */
-/*   Updated: 2023/08/13 19:22:49 by mde-arpe         ###   ########.fr       */
+/*   Updated: 2023/08/14 02:40:40 by mde-arpe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,33 +16,13 @@
 # ifdef MALLOC_DEBUG
 #  include "malloc_debug.h"
 # endif
+
 # include "lexer.h"
+# include "parser.h"
+# include "libft/libft.h"
 # include <stdio.h>
 # include <sys/stat.h>
 # include <readline/readline.h>
 # include <readline/history.h>
-
-typedef enum e_redir_flag {
-	IN,
-	OUT,
-	OUTAPPEND,
-	HERE_DOC
-}	t_redir_flag;
-
-typedef struct s_redir {
-	char			*where;
-	t_redir_flag	type;
-}	t_redir;
-
-typedef struct s_command {
-	char	**args;
-	t_redir	*redirections;
-}	t_command;
-
-// this is the list that the executer will recieve (beta)
-typedef struct s_command_l {
-	t_command			cmd;
-	struct s_command_l	*next;
-}	t_command_l;
 
 #endif
