+++ utils.c
@@ -6,7 +6,7 @@
 /*   By: mde-arpe <mde-arpe@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/29 20:53:05 by mde-arpe          #+#    #+#             */
-/*   Updated: 2023/08/12 18:11:42 by mde-arpe         ###   ########.fr       */
+/*   Updated: 2023/08/14 18:32:21 by mde-arpe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
